requirements.txt
streamlit
pandas
plotly
matplotlib
requests
openai>=1.0.0
python-dotenv